requirements.txt
sentence-transformers
newspaper3k
nltk
fastapi
uvicorn
python-dotenv
faiss-cpu
numpy